Reorganized plugins marketplace folder
