+++ ScriptBeeClient/src/app/plugins/components/plugin-marketplace-dashboard-list-item/plugin-marketplace-dashboard-list-item.component.scss
