+++ ScriptBeeClient/src/app/plugins/components/plugin-marketplace-dashboard-list/plugin-marketplace-dashboard-list.component.scss
