+++ ScriptBeeClient/src/app/plugins/components/plugins-marketplace-dashboard/plugins-marketplace-dashboard.component.spec.ts
@@ -1,12 +1,12 @@
 import { MatListModule } from '@angular/material/list';
 import { PluginsMarketplaceDashboardComponent } from './plugins-marketplace-dashboard.component';
 import { createComponentFactory, createHttpFactory, HttpMethod, Spectator, SpectatorHttp } from '@ngneat/spectator';
-import { PluginService } from '../../services/plugin/plugin.service';
+import { PluginService } from '../../services/plugin.service';
 import { MockComponents, MockDirectives } from 'ng-mocks';
-import { PluginMarketplaceDashboardListComponent } from './plugin-marketplace-dashboard-list/plugin-marketplace-dashboard-list.component';
+import { PluginMarketplaceDashboardListComponent } from '../plugin-marketplace-dashboard-list/plugin-marketplace-dashboard-list.component';
 import { By } from '@angular/platform-browser';
-import { MarketplaceProject } from '../../services/plugin/marketplace-project';
-import { createMarketplacePlugin } from '../../../../test/marketplacePluginUtils';
+import { MarketplaceProject } from '../../services/marketplace-project';
+import { createMarketplacePlugin } from '../../../../../test/marketplacePluginUtils';
 import { MatFormField, MatLabel } from '@angular/material/form-field';
 
 describe('PluginsMarketplaceDashboardComponent', () => {
@@ -16,7 +16,7 @@ describe('PluginsMarketplaceDashboardComponent', () => {
   const createComponent = createComponentFactory({
     component: PluginsMarketplaceDashboardComponent,
     declarations: [...MockComponents(PluginMarketplaceDashboardListComponent, MatFormField), ...MockDirectives(MatLabel)],
-    imports: [MatListModule]
+    imports: [MatListModule],
   });
 
   function mockGetAvailablePluginsApi(
@@ -58,7 +58,7 @@ describe('PluginsMarketplaceDashboardComponent', () => {
     mockGetAvailablePluginsApi(component, [
       createMarketplacePlugin('1', 'plugin-name1'),
       createMarketplacePlugin('2', 'plugin-name2'),
-      createMarketplacePlugin('3', 'plugin-name3')
+      createMarketplacePlugin('3', 'plugin-name3'),
     ]);
 
     const pluginMarketplaceDashboardListComponent = component.debugElement.query(By.directive(PluginMarketplaceDashboardListComponent));
