+++ ScriptBeeClient/src/app/plugins/services/plugin.service.ts
