+++ ScriptBeeClient/src/app/plugins/services/plugin.ts
